backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
celery==5.2.7
redis==4.5.5
httpx==0.24.0
GitPython==3.1.31
docker==6.1.2
pytest==7.3.1
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1 